Improve deprecation message
